Reviewed issues 61, 42, and 86; made some changes
Issue #61
Tweaked :BlkCtr, :BlkCndCtr, :BlkExpCtr masters.
Set spacing for parenleft.case and parenright.case.

Issue #42
The /r looks great.
The /g needs some work that @tphinney is wrapping into issue #100
The /m has interpolation issues between the :LtEtc and :MedEtc masters. This is due to the second (overlapping) node in the :Med weights overlapping with the node one clockwise from the start node, and it should be overlapping with the corner node where the curve will appear in the light. I've fixed this.
The /m start node on the right element is incorrect in the :LtEtc masters (should be one node counterclockwise from the top left most smooth node on the curve). I've fixed this.

Issue #86
For the most part, /v looks good. The :BlkCnd needed some tweaks, so I did that.