Changed "Contrast" axis to "Y Opaque" ('YOPQ' tag), and range changed (and reversed!)
So, the range of the axis was 0 (low contrast) to 100 (high contrast, thin horizontals).

The renames axis is defined in terms of the effect at the default, Medium weight, on the thickness of the horizontal strokes of the caps, per mill (1000 units, unlike the font’s design space of 2000 units). So it ranges from 18 (thin horizontals, high contrast) to 122 (thick horizontals, low contrast).