Updated variable font TTFs
Updated with latest glyph complement and OT features.

Note that the FontLab-exported TTF features a test implementation of FeatureVariations, with two alternate glyphs for /dollar and /Q, which kick in when things get to bold/condensed for good shapes with the default forms. Just need to work out some kinks to get this working in the UFO > FontMake workflow.