Multiple updates:
Moved crossbar of :Blk /h.smcp down 10u. (issue #175)

Adjusted roundness of curves in /s (a little further from the /o curvature and a little closer to the /e curvature) in the :BlkEtc masters. Adjusted upper counter width (no change to stroke width) in the aforementioned masters as well as :Cnd and :CndCtr. Adjusted spacing as well. (issue #158)

Increased tie thickness in the :Blk, :BlkCnd, and :BlkExp masters of the /e.smcp to 312; increased length of top stroke by 5u in the :BlkEtc. (issue #173)

Reduced the /z.smcp :BlkCnd horizontals too 420u to match the :Blk, which worked well. Reduced the horizontals in the :BlkExp more, to 390u to get the color right. (issue #174)

Adjusted spacing of /S in the :Blk, :BlkCnd, and :BlkCndCtr.